requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-filter==22.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
gunicorn==20.1.0
importlib-metadata==4.12.0
Markdown==3.3.7
Pillow==9.2.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
zipp==3.8.0
